+++ src/domain/dimension.js
